Fixed Keyboard not hiding. Moved BCB to first firware ECU. Exceptions (oopsie). Change progress bar colorscheme to match "Renault" colors. Slight visual improvement debug lines (spacing)
