+++ app/src/main/java/lu/fisch/canze/exceptions/NoDecoderException.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package lu.fisch.canze.exeptions;
+package lu.fisch.canze.exceptions;
 
 /**
  *
