+++ app/src/main/res/layout/activity_braking.xml
@@ -131,7 +131,6 @@
             android:text="@string/label_eb_torque_request"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignParentLeft="true"
             android:layout_marginTop="10dp"
             android:textColor="#000"
             />
@@ -202,6 +201,7 @@
                 />
             <TextView
                 android:id="@+id/textDebug"
+                android:paddingLeft="8dp"
                 android:text="@string/default_debug"
                 />
         </TableRow>
