+++ app/src/main/res/layout/activity_charging.xml
@@ -409,6 +409,7 @@
                 />
             <TextView
                 android:id="@+id/textDebug"
+                android:paddingLeft="8dp"
                 android:text="@string/default_debug"
                 />
         </TableRow>
