+++ app/src/main/res/layout/activity_firmware.xml
@@ -274,6 +274,7 @@
                 />
             <TextView
                 android:id="@+id/textDebug"
+                android:paddingLeft="8dp"
                 android:text="@string/default_debug"
                 />
         </TableRow>
